Fix datatype from 'int' to 'long'

+++ check_palindrome.cpp
@@ -1,19 +1,18 @@
 #include <iostream>
 using namespace std;
 
-bool check_palindrome(int n);
+bool check_palindrome(long n);
 
 int main()
 {
-    int number = 99999;
-    int maxj=11111;
+    long number = 99999;
+    long maxj=11111;
 
-    int max = 0;
-    int i,j;
-    
+    long max = 0;
+    long i,j;
+
     for (i=99999; i>11111;i--)
         for (j=i;j>maxj;j--) {
-            
             if (check_palindrome(i*j)) {
                 if (i*j > max) {
                     maxj=j;
@@ -27,10 +26,10 @@ int main()
     return 0;
 }
 
-bool check_palindrome(int num) {
-    int n = num;
-    int rev = 0;
-    int digit;
+bool check_palindrome(long num) {
+    long n = num;
+    long rev = 0;
+    long digit;
     do
      {
          digit = num % 10;
